refactored-router-and-logger
